Major Upgrade PR rewards cleanup (#91)
Very minor PR. Just moves some test code from rewards_test.go to
rewards_v2_test.go, deletes files we're no longer using, then renames
rewards_v2 to just "rewards"
